- Implemented sell ram - Added a query for calculating the current ram price

+++ eos-chain-actions/src/main/kotlin/com/memtrip/eos/chain/actions/transaction/account/BuyRamBytesChain.kt
@@ -34,7 +34,7 @@ class BuyRamBytesChain(chainApi: ChainApi) : ChainTransaction(chainApi) {
 
     data class Args(
         val receiver: String,
-        val quantity: Int
+        val quantity: Long
     )
 
     fun buyRamBytes(
